mess with git lfs
on branches psychophys and office, up until recently, .mat and .log
files (and maybe .png and .csv) were tracked with git-lfs. But the
free data quota was reached.

The present merge is an attempt to stop using git-lfs. Not sure if
the branches office and psychophys will be usable after this...